HtmlUi: Requests must be executed consecutively
It is important that a request cannot overtake another. Example: User
selects a row and clicks a menu. If the selection request is slower than
the menu click, the menu click will probably fail since it needs the
selected row.
Also made sure that the busy cursor does not flicker. Example: User
clicks on a table row, selected and click event are sent by separate
requests. The busy cursor should not be displayed 2 times (for each
request) but only once.
Also fixed offline mode: Request was not queued anymore