HtmlUi: Add splitbox to DOM before rendering child widgets
This seems to fix "unspecified error" bug in IE when manipulating
"selectionStart" property of a string field while rendering it. Error is
kind of obscure and cannot be easily reproduced with a simple example
(but can be in a very complicated one, i.e. wizard). Seems to have to do
with the string field not yet fully inserted in the DOM tree.

#163205