HtmlUi: Prevent "EOF exceptions" when reloading the page
When UiSession is disposed, background jobs must be notified immediately
(without 100ms delay). Otherwise, the client already disconnected and
writing to the output stream will cause an exception.

Because writing to a closed connection cannot be prevented for sure, we
catch the resulting EOFException and write a shorter warning to the log.