HtmlUi: Add "reference" to JsonEvent to remove them as well on dispose
This fixes errors when opening and closing a form in the same request.
Because "showForm" and "hideForm" are targeted to the desktop (which is
not disposed) they would not be removed when the form is disposed. This
causes errors on the UI because the form is not available anymore. To
fix this, there events are marked with the form as "reference", to be
able to remove them as well wenn removeJsonAdapter() is called.