HtmlUi: Add very (!) basic implementation of "tile"
Adjustments required for tiles:

>>> Widget.js:
Because Tile.js is a model adapter but not really a widget (has no own
representation in the DOM), it provides no key stroke context (returns
null when _createKeyStrokeContext() is called). Widget.js had to be
adjusted to allow return value of null.

>>> LogicalGridLayout.js:
LogicalGridLayout did not respect CSS margins of components. Scout
fields did not have a margin anyway, but tiles use a margin, therefore
the LGL had to be fixed.

>>> FormFieldLayout.js:
Calculation of field container size and position was fixed. When
top/left are adjusted, the available sizes decreases. Also, when a
padding is configured on the container, the position of the field has to
be adjusted.