HtmlUi: Outline: Move layout validation to postRender() to fix dialogs
When dialogs are rendered, the outline layout has to be valid.
Otherwise, the correct document height cannot be retrieved.

This is relevant when the users presses Ctrl+R while an outline is
showing whose tree content is higher than the browser window and a
dialog is open whose content larger than the browser window.