JSONObject: Fix return value of get/optString for "null" values
The JSON property value "null" is mapped to the sentinel object
JSONObject.NULL. When retrieving such a value with getString or
optString, the original Android implementation returned the literal
string "null", which is not correct. Instead, Java's null value has to
be returned.

A test was added to check the behavior of the get/opt methods for all
supported types in case of null or undefined property values. Because
the original org.json implementation already had the correct behavior,
the test was named "CompatibilityTest".