TreeEventBuffer: Improve coalesce of same consecutive event types
When same consecutive node event types are coalesced, the common parent
node (CPN) has to be checked as well. However, within a group of same
event types, the same CPN does _not_ have to be consecutive.

Example:
1-CHECKED[A], 2-UDPATED[B], 3-CHECKED[C], 4-CHECKED[D], 5-CHECKED[C],
6-CHECKED[A]

Events #1 and #6 cannot be merged, because #2 is a different event type.
However, event #3 and #5 may be merged, although #4 has a different CPN
(but the same event type).