HtmlUi: Popup close handling improved
Requirement:
Form popup (e.g. TableOrganizeForm) needs to be able to open other
popups (context menu, cell editor) or dialogs without beeing closed when
these popup are opened or these popups/dialogs or theirs content get
clicked by the user

Solution:
In order to detect whether the mouse click happened outside of the
popup, we do not check the dom hierarchy anymore but the widget
hierarchy. Example: The dom parent of a context menu is the scout div,
the widget parent of a context menu is a table, tree or another widget.

Requirement:
The widget hierarchy needs to be correctly set up.