HtmlUi: Handle ?poll response after user response
When a user request is pending on the server and a ?poll response
arrives as the UI, the ?poll response is now queued until the user
response is handled. This prevents some problems with asynchronous
events arriving to early on the UI.

Example:
1. ?poll request starts and waits on server [R1]
2. User clicks a button
3. User request [R2] in handled on server in a model job, a new form is
created
4. Model job schedules a client job to update some data on the new form
5. User request [R2] returns and releases the UI session lock
6. Async job starts model job to update the form. It deletes some rows
in the new form. Because the user request has already left the
UiSession, the polling job is awaken. [R1]
7. Before the response of the user request [R2] arrives at the UI, the
background polling job returns to the UI [R1]. This might happen because
this response is much smaller.
8. The UI handles the "row deleted" event. This throws an error because
the form is not yet created.

Although the UiSession and the jobs on the server are synchronized, the
requests might overtake each other because the network between the UI
and the server is unreliable and there a no guarantees about the
sequence of events.

This commit's solution does not solve all problems, but the remaining
ones should be even more uncommon.

#164326