HtmlUi: Prevent initial focus in tile container box
Tile fields are not focusable, but their content may be (e.g.
app-links). To prevent automatic scrolling when the tiles are rendered
and the first app-link is off-screen, the FocusManager is extended to
not focus elements that have the class '.prevent-initial-focus'. The
tiles container box has this class assigned by default.