HtmlUi: Make table footer responsive
Table footer now has a compactStyle which gets active if there is not
enough space to display the whole content.
If this style is active, the info section gets displayed in a compact
mode. Additonally (not implemented yet) The table controls should
collapse as the menus in a menubar do.

Also refactored table footer and css classes to conform with current
coding guidelines.