HtmlUi: Tree: Render selection in _postRender()
If the selected node is not visible initially (because one of its
parents is not expanded), all parents are expanded recursively to make
the node rendered. If this would be done inside _render(), the property
"rendered" would still be false. This causes the expanded parent nodes
not to be rendered, which in turn breaks the rendering of the selection.
Therefore, the selection should not be rendered until the initial tree
rendering is complete.