HtmlUi: Fix table control layout when layouted in background
TableFooter has a layout, it is not necessary anymore to call
footer.onResize() manually. This method was therefore deleted. Instead,
the control is resized from the table footer layout.

When the table control is open and on a form that is in the background,
the size of the open content has to be subtracted from the table data's
height. To calculate this height, the "height: 0" rule has to be removed
from the CSS. It is added manually while opening the container.