HtmlUi: Improve layout behavior of Graph and Chart on session init
When the graph or chart are on a form in the background and the page is
reloaded, the layout used to be broken because of the debounce delay.
When the layout is started, the form is detached and the correct hight
cannot be determined anymore. Therefore, the debounce must _not_ occur
when the desktop is initially rendered. This can be checked by looking
at the new property "session.ready".

Session.js:
this._startup and this._unload were removed and replaced by specific
_sendStartupRequest() and sendUnloadRequest() methods. _sendNow() is now
responsible for queued async events only. Additionally, two new flags
were introducted: session.ready = true after the desktop has been
completely rendered after the initialized event. session.unloaded = true
after the unload event has been received from the window.

ChartTestForm.java:
Legend position radio group box was changed to smart field to make the
form look better on small screens. Some NPEs were fixed. The code
formatting was fixed (somehow, the indent size was broken).