AbstractWizard: Changed call of decorateWizardContainerForm()
The argument 'containerForm' was removed from
decorateWizardContainerForm(). It can be retrieved with
getContainerForm().

The decorate method used to be called within
execCreateWizardContainerForm(). This place is semantically and
technically wrong. Because the container form instance is built during
initConfig() (= constructor) the getConfigured* properties of the wizard
would not be applied. Therefore, the decoration was moved from the
creation to the start() method.