removed unnecessary placeholder index declaration in log messages
In LOG.info("msg {0}", obj), '{0}' is replaced with the string
representation of obj. However, the index within the braces was not used
at all. Hence the following statements result in the same logged
message:

LOG.info("{0} {1}", a, b)
LOG.info("{1} {0}", a, b)
LOG.info("{} {}", a, b)

Scout no-OSGi is switching to slf4j which does not support indexed
placeholders at all.

Change-Id: I7fa7fb54857b3057e572166af640c1203bbf96d1
Reviewed-on: https://git.eclipse.org/r/59715
Tested-by: Hudson CI
Reviewed-by: Andi Bur <andi.bur@gmail.com>