IWizard: Remove confusing properties, fix container form decoration
Former wizard properties "displayHint", "displayViewId", "modal" were
removed, because they should be applied to the wizard container form. If
the wizard would do that automatically, it might overwrite values of the
container form. Therefore, the programmer has to set the properteis
explicitly if he wants so. For that, the method
"execDecorateContainerForm" is provided. The form should be created in
"execCreateContainerForm". The container form must not be null.

Former wizard properties "iconId", "tooltipText" and "wizardNo" were
removed because they were never used.

Property "titleHtml" was removed because it is a legacy property.