update to MM 1.6.9
jaxws-modules must define their dependencies explicitly, they don't receive them via automatic profile activation.
Dependencies from profiles are obscure and not immediately visible if the pom. Only if the effective pom is presented all dependencies are visible.

(Second it doesn't work with the flatten-maven-plugin as it strips the profiles from the resulting pom. So no dependency information must be there.)

Change-Id: Ie00da2cca7e0563a2f3cb2e1e90ee025af3073d2
Reviewed-on: https://git.eclipse.org/r/60530
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsi-software.com>
Tested-by: Stephan Leicht Vogt <stephan.leichtvogt@bsi-software.com>