Remove ServerJobs and ClientJobs as obsolete now
ServerJobs and ClientJobs are not necessary anymore, because simply working on another RunContext. In the very beginning of the Job API where the concept of RunContext did not exist yet, that semantic was incorporated within the jobs.

Also, the RunContext must be provided explicitly and is not resolved based on the current calling context anymore. That was changed to standardize the job's usage and to explicitly think about the context in which to run a job.

Also, simplified Future and JobEvent based filters by using the builder pattern.