AbstractColumn: Don't apply user prefs when header is disabled
When the header is not enabled, only the configured values should be
applied to the column, because the user would not be able to change or
reset the settings.

Additionally, if a column has fixed with, any previously stored width
from the user preferences should be ignored (for the same reason).