HtmlUi: Use JSONObject.NULL instead of "" for null values
The JsonResponse would not emit properties with a null value (because
that is considered as "not present" a.k.a. "undefined"). To explicitly
send null as value in a JsonResponse, the special marker object
JsonResponse.NULL has to be used. This used to be an empty string, but
this is not correct for non-string data types.