Release threads waiting for a job to complete after transition into 'done' state
The 'done' parking mechanism of Java {@link FutureTask} cannot be used by {@link JobManager}, because waiting threads are released before {@link FutureTask#done()} is invoked. That results in a inconsistent state, because bookkeeping and event firing is to be done prior releasing any waiting thread.

This change also fixes the regression failures in job JUnit tests.

Change-Id: I7e54a23600b98eceaf9c1d684596134182cfc9c7
Reviewed-on: https://git.eclipse.org/r/61235
Tested-by: Hudson CI
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsi-software.com>