HtmlUi: Reorder some jquery-scout functions, fix $.fn.icon()
Reorder:
All "make" functions were moved together to the top of the file. The
makeSVG() function was placed near the other SVG functions.

Fix $.fn.icon()
The icon function did not work correctly when a font icon was changed to
a picture icon and vice-versa. A jasmine test was added for this.