Fix job regression related to thread interruption
If a thread is waiting for a condition to fall and gets interrupted, that thread does not immediately exit with an InterruptedException. That is because a thread's interruption is asynchronous, meaning that once a thread (a) interrupts another threads (b), and in turn thread (a) unblocks the condition which thread (b) is waiting for, it is not ensured that thread (b) exists with an {@link InterruptedException}. However, many tests expect exactly that behavior.

Change-Id: I07b2f72e5b0a62d4d74a8f00af6bae8f96c03077
Reviewed-on: https://git.eclipse.org/r/61310
Tested-by: Hudson CI
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsi-software.com>