HtmlUi: Refactoring menubar and popups
Remove 'line-height' style alignment from menu-items. Instead, content
should always be centered vertically. This is implemented as follows:
- ::before or ::after element with height=100% spans the container. This
is available as mixin: #scout.vertical-align-helper-before/after().
- Items to be centered are set to display: inline or inline-block. They
are assigend the "vertical-align: middle" class.

Positioning of popups and menubars had to be changed accordingly.

Basically, this change allows to use icons in menus.