HtmlUi: Use non native tooltip for table cell
Additional to the non native tooltips, this change includes a fix for
safari: Safari automatically displays a tooltip if the text is
truncated. Because we add a js based tooltips in certain cases, two
tooltips may be shown at the same time. This looks bad, so current
workaround is to never install js based tooltips if the text may get
truncated, if safari is used. Other browsers still show the styled js
tooltip.