HtmlUi: Table virtual scrolling: Update row height and viewport on column resize
When having a column with wrap text, the scrollHeight may change after
resizing a column -> fillers and viewport needs to be adjusted.
Also, if text of a cell with wrap text changes, the cached row height
needs to be recalculated.

Change in HtmlComponent: Check for isAttached is necessary because
height of table rows cannot be calculated if table is not attached.
Actually it is not necessary at all to call setSize/setBounds for not
attached components because pref size cannot be calculated -> saves
unnecessary layout operations.