JAX-WS: Fix authentication issue in one-way communication
Problem:
JAX-WS Metro RI v2.2.10 does not exit the call chain if the ongoing request is a one-way communication and AuthHandler returns with 'false'. That results in that the endpoint operation is still invoked.

Solution:
Throw HTTPException with respective HTTP status code instead.

Note:
This issue was already addressed in JAX-WS Scout with Metro RI v2.1.6, but the support to identify one-way-operations was removed in a later Metro version. [currentVersion=v.2.2.10]

Change-Id: I2c6c3423335294f212bbf3fbf34efff39acd87f3
Reviewed-on: https://git.eclipse.org/r/62836
Tested-by: Hudson CI
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsi-software.com>