HtmlUi: Fix display of selected toggle buttons in menu bar
Use _setProperty instead of simply setting the property value. This will
trigger an event. Otherwise, the event would not be triggered until the
response from the server is received, and that event would have no
effect, because the 'selected' property was already changed.

Additionally, Menu.js must call super._renderSelected() to add the
'selected' class to the $container.