HtmlUi: Improve error handling in JsonResponse
When JsonResponse.toJson() throws an error, a new response object should
be created (because the response's content is apparently corrupt
somehow). To prevent timeouts in the response queue due to the missing
response, the original sequence no. is written to the UI error.

Removed assignSequenceNo() from JsonResponse, instead the sequenceNo
must be passed to the constructor.