HtmlUi: Remove partly wrong and confusing "grouping" check
Correct groupings cannot be checked that easily. Furthermore, the Java
DecimalFormat accepts an input like "2'5" (with ' as grouping character)
and parses it as "25". It is therefore futile for JavaScript to check
for correct groupings. Even if the method would return when the grouping
is wrong, the Java logic would automatically "fix" it.

#164389