ContentAssistField: Join model synchronously when not in model thread
When the result of a lookup fetch is applied to the field, this must be
done inside a model thread (because the model is changed). If the
current thread is not the model thread (i.e. fetching happened
asynchronously), a model thread is scheduled. After that, the current
thread must wait for that model job to be done, otherwise the
LookupRowCollector might throw an exception, because the result it not
yet written to the collector ("Lookup row fetching not completed yet").