Exclude "batik-js" from Maven dependencies
batik-js contains a patched version of Mozilla Rhino, see
http://svn.apache.org/repos/asf/xmlgraphics/batik/trunk/lib/README.js.txt.

This dependency conflicts with official Rhino classes on the class path
of a Scout application. To fix this, batik's dependency is excluded.
(The patch is very old and not needed by Scout anyway, because the Scout
Swing UI is now obsolete.)

The batik-swing dependency (or the dependency to the entire batik
package) cannot be removed yet, but will be probably at some point in
the future.