HtmlUi: Fix positioning of large popup dialogs
When a dialog is opened, the DialogLayout ensures that the dialog bounds
do not exceed the browser window. To get the max. size, the $(window)
object must be used, because $(document) might be wrong, when the
dialog's preferred size is very large.

Additionally, the dialog is centered on screen. When the dialog is
moved, popups (such as tooltips) were not repositioned. This was fixed
by introducing a 'move' event on Form.js. Tooltips subscribe to this
event to position themselves correctly.