Cancel job before canceling the associated RunMonitor
This is crucial, as otherwise the interrupted thread may terminate the task's execution even if the job is cancelled. That is because interruption is done by the context's monitor.

However, Java FutureTask cancels the job without interrupting its executing thread (even if requested). That is done later by RunContext, or JobFutureTask if not specified. That ensures proper monitor and future state once the thread is finally interrupted.

Change-Id: I8f00d72e2d8e7778ea65c01ee49288d7b32cd041
Reviewed-on: https://git.eclipse.org/r/64083
Tested-by: Hudson CI
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsi-software.com>