Fix issue that sometimes Smartfield lookup does not work as expected
ContentAssistFieldUIFacade.proposalTypedFromUI(text) rejects requests if the entered text was already fetched. This behavior is as expected.
However, the current fetch-result was not reset before starting a new lookup search. So if typing fast, and there is an asynchronous fetch operation already in progress, then a new lookup search was rejected if its search-text is the same as the search-text before the operation still in progress. As a result, the lookup rows displayed do not match the entered search text.

Pairing: Claudio.Guglielmo@bsi-software.com
Change-Id: I06280b8f9c09be2714c52d205b10eaed59e35920
Reviewed-on: https://git.eclipse.org/r/64287
Tested-by: Hudson CI
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsi-software.com>