Respect the semantic of asynchronous 'LookupRow' fetching even for LocalLookupCall
There are consumers using asynchronous data fetching an expect data to be loaded asynchronously.
In a previous version, this was ensured by a redundant model job in AbstractContentAssistField.fetchLookupRows(...) when writing data back into the model. However, this is the wrong approach, because asynchronisity refers to how data is fetched, and not how data it is written into the model.

This fixes the problem of selecting the current row in TreeProposalChooser.
          
Pairing: Claudio.Guglielmo@bsi-software.com
Change-Id: If579dee0db547fc746d02111bf5202e5a6e22183
Reviewed-on: https://git.eclipse.org/r/64286
Tested-by: Hudson CI
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsi-software.com>