ActivateForm does not work for views with displayParent outline
Since activateOutline fires an event to switch to the outline or, if the
outline already is selected, bring it to the front. The second behavior
prevents a form from being activated, because every time the tab gets
clicked activateForm is called which calls activateOutline -> Tab gets
deselected right after getting selected.
-> Do not activate outline if it already is selected