HtmlUi: Virtual scrolling in editable tables should consider popups and tooltips
If a cell editor is open and the table gets scrolled, cell editor needs
to be removed if the row is removed, and rendered if the row gets
rendered. Along with this the popup of the cell editor (e.g. datepicker)
should not be visible if the cell editor is out of view.