HtmlUi: Fix date picker popup in Firefox
Because "-moz-user-select: -moz-none" is no longer automatically set on
all DIVs, the date picker popup was broken in Firefox. The date field
got a "focus lost" event which caused the picker to be closed before the
value could be selected.

As a fix, the focusUtils.isSelectableText() method was changed to mimic
the "inheritance behavior" of the -moz-user-select property. As a
consequence, the exception rule in main.css could be reduced.

#168809