HtmlUi: Use UIFacade when checking/unchecking a check box from the UI
When accessing the model directly from the JSON layer, the client run
context is not applied correctly, which leads to certain ThreadLocals
not set (e.g. IDesktop.CURRENT.get()). Instead, the JSON layer must
always call the methods on the UIFacade, which are proxied by the
ModelContextProxy.

Additionally, the UIFacade usually checks for isEnabled() and
isVisible(). Without that check, an (evil) UI could change values it
should not be able to.