Fix timeout while waiting for other jobs during session shutdown
awaitDone() now throws a TimedOutException when the jobs did not
complete within the specified duration (this used to be a boolean return
value). When this happens, the call to cancelRunningJobs() was skipped.
Instead of printing the running job futures to the log, an error message
was logged.

The method was refactored in the following way:
- There is only one method: cancelRunningJobs()
- The method always cancels other jobs, but waits only when the config
value is set.
- The filter was simplified (no need to filter for ClientRunContext,
because without ClientRunContext, the session will not match either) and
is only built once.