HtmlUi: Smartfield and tableheader menu have wrong width
The popup layout tries to adjust the size if the popup does not fit into
screen. In case of regular popups (smartfields, tablehader, etc) this is
fine for the height but it does not work for the width. Reason: The
popup is getting moved if it overlaps the window border compared to a
popup with head (menu, form tool button) which are switched to the other
side. This has the effect that the popup will be too small if the space
left and right of the anchor is small.