Unset 'wait-for' hints immediately upon being unblocked
This guarantees the hints to be reflected immediately, so they can be
evaluated by the caller. Otherwise, they would be unset asynchronously
upon wake up of the blocked threads.

Use case: 
A 'blocked' Form gets unblocked via a JSON request. Upon unblocking the
Form, that JSON request likes to wait for all running model jobs, which
do not require UI interaction. That requires the hint to be removed
immediately, as the request would not wait for the released model job
otherwise.

Change-Id: I42a080dc5b16f1597aea56036a833255a9729381
Reviewed-on: https://git.eclipse.org/r/66082
Tested-by: Hudson CI
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsi-software.com>