HtmlUi: Pages visible but rows are not (filter)
If table row has an active filter, the pages should be filtered as well.
This doesn't work always, rows may get added but no filtereChanged event
fired -> scout.DetailTableTreeFilter is never executed again and node
has filterAccepted = true because link to row was not established yet.
Happened in global search when same search was executed while a filter
was active.

Btw: The added test could not reproduce the bug, it is an outline /
table combination issue.