HtmlUi: Use dedicated startup response when initializing UI session
UiSession.java:
Replace EVENT_INITIALIZED by a special JSON response with StartupData
object. It is used to transport the assigned uiSessionId to the UI. The
old event was not practical for that purpose, because the event handling
on the UI cannot find the event target (because the uiSessionId is not
yet known).

Session.js:
Session startup params are no longer stored as member on the session
object. The initial startup request is handled by a dedicated method and
does not use the "normal" event processing mechanism. This is necessary
to assign the uiSessionId and register the session in the model adapter
registry.

URL.js:
Made "raw" fields public.