HtmlUi: Add possibility to disable lazyExpanding of the tree
LazyExpandingEnabled may be set for each node individually which is
fine. In order to disable it, it would be necessary to traverse the
whole tree and listen for insert events only to set the flag to false.
To make it easier, the tree should provide a property to disable the
feature for the whole tree.