HtmlUi: Add layouts for desktop, navigation, bench and table controls
Desktop needs to be more flexible. A first step is to have layouts to
make it controllable using javascript.
Additionally, root components like form, table, tree, table controls
have a onResize function which is called when the desktop gets resized.
This is different to every other widget which only work using layout ->
Remove that onResize to be more consistent and flexible.