HtmlUi: LayoutValidator should validate even if there is no request
LayoutValidator now uses setTimeout to schedule the validate operation.
Unfortunately, the browser may execute rendering before executing our
scheduled task, which has the effect that the layouting is visible to
the user, especially for large forms. That means we still need to call
validate at some points explicitly.

Scheduling a microtask would be the solution, but does not seem to be
possible yet
(https://jakearchibald.com/2015/tasks-microtasks-queues-and-schedules/)