HtmlUi: Navigation resizing flickers
If splitter is grabbed slightly right from the center and pulled left
while outline has the breadcrumb style, the tree is visible for a very
short time before style is set to breadcrumb again. This does not happen
if the center of the splitter is grabbed.

Fixed by determing the cursor offset on mousedown.