HtmlUi: Add "sync" request to fetch missed JSON responses from history
All JSON responses are stored the UI session. With each request, the UI
sends the sequence no. of the last response it had successfully
processed. That response (and all that are older) is then removed from
the history again.

When the UI comes online after it has been offline, it sends a special
"sync" event to the server to fetch all entries from the JSON response
history. Any response that has been created while the UI was offline is
then processed.