HtmlUi: Don't use jQuery object directly
Scout uses an IIFE to wrap the code in a function. Inside this function
(-> inside the scout module) the variables $ and scout are available.
This makes it easier to include other libraries which also work with $
or even include other jQuery versions. The only premise is that we never
work with jQuery directly, only with the $ variable.