Fixed problem with table-event-buffer causing UI errors.
In complex cases with filtered rows the UI sometimes received too much
or too few rows, which caused a javascript error in the UI. Most of the
bugfix is about the row-order-changed event, which requires the UI
(browser) to have exactly the same row-count as the model. When these
two numbers are not the same an error is thrown in
Table.js#_onRowOrderChanged. 

See ticket #170668