Add malware scanner to file uploads
All file uploads in scout are processed in the UploadRequestHandler.
This change adds the new @Bean MalwareScanner to the scout platform
enabling malware scan of all uploaded files using the in-place installed
malware scanner of the target server.
The file to be scanned is copied into the temp folder of the OS and
assumes that the installed malware scanner has realtime filesystem check
active and therefore removes unsafe files immediately.

Change-Id: Ibe7d8bbaaeec38ba157a88fbbb9a313cbc1dd2eb
Signed-off-by: Ivan Motsch <ivan.motsch@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/67903
Reviewed-by: Judith Gull <jgu@bsiag.com>
Tested-by: Hudson CI