LogicalGridLayout: Fields with gridh=1 and weighty=0 must never scale
UseCase: ListBox with gridh=5 and top label ond the left. On the right a
placeholder field with gridh=4 and a field with gridh=1.
The current behavior is that the field on the right side is bigger than
the logical grid row height. This is wrong, the placeholder should grab
all the space.
Reason: The top label on the left side increases the size a bit and
makes all cells bigger.
Solution: Only scale cells with weighty > 0. This was already done for
the columns (initializeColumns) -> use the same logic for rows