Mobile: Create table row overview for nodes without a detail form
When browsing in mobile mode and a node gets selected, the detail form
is embedded. If there is no detail form we should provide a summary of
the table row -> TableRowDetail.js
Furthermore, this browsing style requires that table pages always have
child nodes, otherwise it is not possible to show the table rows -> Add
new property alwaysCreateChildNode and set it to true in mobile mode.