Desktop: Separate view tab bar from header
Extracted view tabs into separate file so that it can be reused and is
not bound to the header.
Also removed view rendering logic from DesktopViewTab, this class should
only be responsible for the tab -> Necessary if header can be made
invisible.
Also fixed displayParent logic if display parent is outline and there is
no navigation -> dialog could not be opened if outline is not rendered
due to invisible navigation, this is not expected.