Table: Remove rowHeightHint
This property was added because with rap and swt it was not possible to
have variable table rows as height as their content. The only
possibility to get multiline rows was to set a fixed height. This
limitation is now gone. If you still want every row to have a fixed
height on multiline tables, you can use css to achieve it.