Popup: Rework mouse down handling
A popup attaches a mouse down listener to close the popup when clicking
outside. Since the popup is most often opened with a mouse down as well,
attaching a mouse down listener while mouse down is executed would
prevent opening the popup at all. Thats why a detection was implemented
whether the mouse down event is the same as the one which opened the
popup. In order to do this, event parameter was added to a lot of
methods, which actually destroyed the api. Additionally, it did not work
for every case, e.g. if the popup was opened indirectly using another
widget.
To fix this, the mouse down listener is now added to the capturing phase
which means it is executed before any other regular (bubble up) mouse
down listener. 

The fix in the menubar prevents layouting the menubar on each property
change (e.g. selection change). Use case: Open a menu in the menu-bar,
then click on the ellipsis menu -> Ellipsis menu was immediately closed
because the menubar was layouted. 

The code in action.prepareAction is not needed anymore because
acceptInput is executed by the keystroke itself (already implemented a
long time ago)