Popup: Fix keystroke handling
My last change destroyed the keystroke handling, which means popups
could not be opened with a keystroke anymore.
To fix this the generic logic to close the popup when any keystroke was
pressed has been removed. The cases when the popups should be closed are
well known, so it is not necessary to do it always.