Desktop: Improve animation
Animation to show/hide the navigation is very slow if the bench shows a
form.
Fixed by removing the relayouting on every step. It has a little
drawback that when showing the navigation a white area on the right side
of the screen is shown. But this is still better than a laggy animation.