Mobile: Add touch highlight
On most mobile devices (surface, android) the active state is shown when
an element is touched. Therefore the clickable elements should have an
active state defined even though it mostly has no effect on the desktop,
because there is a hover state as well.
The native touch highlight on webkit browsers (safari) has been disabled
in favor of the active state.