Mobile: Add touch property to user agent
Nowadays touch is not limited to mobile devices anymore. On the gui we
already know the property 'supportsTouch' but on java side we don't have
this granularity.
The smart field and date fields are opened in a touch friendly popup if
touch is supported. If that is the case, the proposal field needs to run
in a different mode (auto closing the chooser must not happen). In order
to set this mode, we need to know whether touch is true.

In addition to the property touch and the smart field property
autoCloseChooser, we needed a way to set the property when the device is
running in touch mode. Therefore the device transformation handling has
been adjusted in a way that not only one transformer is used but as many
as needed. Example: If it is running on a smartphone, Mobile and Touch
Transformer are active.