Mobile: Make proposal smartfield work
The custom text will be accepted, when the user touches outside of the
smartfield, or if the user presses enter on the virtual keyboard.
Pressing enter will close the popup, pressing done just blurs the field.
This also fixes the issue that the original smartfield never displayed
the error message, which is the case if the user enters invalid text
into the touch popup smartfield.