Menu: Open popup even if menu is inside ellipsis
If menu is overflown (-> ellipsis menu is shown) and gets selected, the
popup needs to be opened. If the popup is attached to the menu with the
header, the ellipsis menu needs to be opened. 
In case of mobile, where the popup is not attached to the menu, the
ellipsis menu should not be opened, it is only necessary to show the
popup.