Mobile: Fix exception when removing bench animated
Currently the bench and header are removed by creating a clone. The idea
was to ignore following hideForm and destroy events which would remove
the form while the bench is being removed by an animation.
This does not work always: When an outline changes while the bench gets
removed (which is likely after a new form was stored), an exception
occurs because the clone doesn't have the required methods.

Now, the bench and header are removed by css animation. When the
animation ends, the actual removal will be triggered. While the
animation is in progress, every attempt to remove will be blocked. This
fixes the exception because no clone is necessary. Another advantage: It
is faster due to the css animation and missing clone.