Bugfix for outline-change with wrong active form.
Happens during deep-link activation. We have an active form X which has
outline A as display parent. The deep-link activates outline B but
active form is still X. This causes an UI error because outline A is not
initialized, only outline B is initialized. Thus the adapter for form X
is never registered -> error.