Page detail content: Support wrapped text
Title as well as the table row detail may contain wrapped text.
Calculating the pref height for these elements is only possible with the
correct width -> Enhanced graphics.prefSize to give a width hint. The
deleted of the white-space no-wrap code is safe because it didn't worked
anyway (would be nowrap instead of no-wrap).
The solution is not perfect, it works only if the container width is
set. If the prefSize is asked by another parent (e.g. desktop
navigation) it would be wrong because the container size is not set.
Maybe someone else has a better idea in the future. For now it works.