Desktop navigation: Shadow vanishes after toggling navigation
If the navigation is hidden and the user presses the key to make it
visible twice (to skip the breadcrumb style) the bench shadow vanishes.
Reason: With every key press a new animation is started. The jQuery
animations set the overflow style of the element to hidden and reset it
back when they finish. If multiple animations are running, the reset
does not work and overflow will be hidden afterwards -> missing shadow.

There is no need to have concurrent animations -> stop the running when
a new one gets scheduled.