Mobile: Menu on node page not visible
If a menu is added to a node it will be wrapped and added to the detail
table as well. Such menus are not visible because they are marked as
outline wrapper and therefore removed by the outline on ui side. The
intention of this removal code is to remove menus which are declared on
the parent detail table and moved to the child page and eventually to
the child detail table using outline wrappers. Actually only such menus
should be removed but this is not detectable anymore on ui side. Because
the code in AbstractOutline only creates wrappers for single selection
menus if the parent page is a table page, it is safe to remove the
outline wrapper menus only in that case.