Mobile: Remove unused code
MobileTable: Although the table is not yet responsive, the approach to
wrap a mobile table around the actual table won't be used. Instead it is
more likely that the ui table will adjust its style when used on small
devices.

DeviceTransformationHook: The hooks are not required anymore, the same
can be accomplished using custom transformers.